Add note about `ItemBlocks` in blocks/blocks.md (#57)
* Add note about `ItemBlocks` in blocks/blocks.md

Add note about registering `ItemBlock`s as well as registering `Block`s,
and note that `GameRegistry.registerWithItem` should not be used.

* Remove warning on `GameRegistry.registerWithItem`

The method is deprecated and the Javadoc notes what to do instead.

As per williewillus's suggestion.

* Explain what `ItemBlock` is in more detail

Add paragraph on the differences between `Item` and `Block`, and how a
block has a representation for each. Describe how `ItemBlock` facilitates
this.