Reduce confusion regarding gradle wrapper scripts
There is no `gradlew.sh`, at least not in the latest mdk (and AFAIK the gradle wrapper script for *nix never had a file extension), the files the docs refer to here are `gradlew.bat` (for Windows) and simply `gradlew` (for anybody else). I think it reduces friction for people unfamiliar with gradle to just explicitly spell out the names of both files and not mention a `.sh` file name extension that does not exist.