1.20.x: correct the spelling of one word within events.md
The word "busses" is invalid since the plural form of "bus" is only
"buses", "busses" should be the third-person singular of "bus" but not the
plural form of it. Consequently correct the mistake.

Within the 1.19.x version the issue above has been tackled, but the one
within the 1.20.x version has not been modified yet. Hence carry out the
fix as well for the latter.

Signed-off-by: src_resources <src.res.211@gmail.com>