1.19.x: improve an expression within items/index.md
The expression "An item(s)" is invalid on grammar when it refers to the
plural meaning at which it turns into "An items". To avoid the possibly
vague thoughts for some readers about this, change this into "An item or
multiple items", towards which readers are able to comprehend both
grammatically and misunderstanding-avoidingly.

Signed-off-by: src_resources <src.res.211@gmail.com>